Replace int64 with int32 everywhere
Using int32 uniformly for all int tensors. int32 seems to give a
non-negligible latency improvement over int64, especially for the
bucket fill operations. int32 is expected to be sufficient for
current use cases. Should be changed if required in the future.

Note: int64 is the PyTorch standard for all indexes.